backend/requirements.txt
fastapi
uvicorn[standard]
pyjwt
pymongo
pydantic-settings
passlib
python-dotenv
